- Jorrit fixed a size_t related mismatch in csEvent and iEvent which 	  caused compile errors on MacOS/X.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@18565 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ include/csutil/csevent.h
@@ -260,7 +260,8 @@ public:
   virtual csEventError Retrieve (const char *name, float &v) const;
   virtual csEventError Retrieve (const char *name, double &v) const;
   virtual csEventError Retrieve (const char *name, const char *&v) const;
-  virtual csEventError Retrieve (const char *name, const void *&v, uint32 &size) const;
+  virtual csEventError Retrieve (const char *name, const void *&v,
+  	size_t &size) const;
   virtual csEventError Retrieve (const char *name, bool &v) const;
   virtual csEventError Retrieve (const char *name, csRef<iEvent> &v) const;
   virtual csEventError Retrieve (const char *name, csRef<iBase> &v) const;
